requirements.txt
asgiref==3.5.1
Django==4.0.4
django-rename-app==0.1.3
djangorestframework==3.13.1
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1